+++ common/service/k8s/appgroup/groupapi.go
@@ -90,11 +90,17 @@ func (a *AppGroupApi) GetAppGroup(namespace string, name string) (*appv1.AppGrou
 }
 
 func (a *AppGroupApi) UpdateAppGroup(namespace string, group *appv1.AppGroup) (*appv1.AppGroup, error) {
+	if err := SyncDependencyLabels(group); err != nil {
+		return nil, err
+	}
 	a.filterEmpty(group)
 	return a.clientset.AppgroupV1alpha1().AppGroups(namespace).Update(a.sdk.Ctx, group, metav1.UpdateOptions{})
 }
 
 func (a *AppGroupApi) CreateGroup(namespace string, group *appv1.AppGroup) (*appv1.AppGroup, error) {
+	if err := SyncDependencyLabels(group); err != nil {
+		return nil, err
+	}
 	a.filterEmpty(group)
 	return a.clientset.AppgroupV1alpha1().AppGroups(namespace).Create(a.sdk.Ctx, group, metav1.CreateOptions{})
 }